+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `Array#key_map` and `Array#dig_map` for mapping over `Hash`
+
 ### Changed
 
 ### Removed